move translations out
