+++ InvenTree/web/src/src/components/nav/Header.tsx
@@ -9,7 +9,7 @@ import {
 } from '@tabler/icons';
 import { useState } from 'react';
 import { Link, useNavigate, useParams } from 'react-router-dom';
-import { languages } from '../../App';
+import { languages } from "../../translation";
 import { useApiState } from '../../context/ApiState';
 import { useLocalState } from '../../context/LocalState';
 import { tabs } from '../../defaults';
